+++ busTrackerApi/src/main/kotlin/busTrackerApi/db/models/Route.kt
@@ -1,0 +1,8 @@
+package busTrackerApi.db.models
+
+data class Route(
+    val fullLineCode: String,
+    val simpleLineCode: String,
+    val routeName: String,
+    val codMode: String
+)