+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/VehicleLocations.kt
@@ -1,0 +1,9 @@
+package busTrackerApi.routing.lines
+
+import crtm.soap.VehicleLocation
+
+data class VehicleLocations(
+    val locations: List<VehicleLocation>,
+    val codMode: Int,
+    val lineCode: String
+)