+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/bus/BusLinesClient.kt
@@ -2,20 +2,22 @@ package busTrackerApi.routing.lines.bus
 
 import arrow.core.continuations.either
 import busTrackerApi.db.getItineraryByFullLineCode
+import busTrackerApi.db.getRoute
 import busTrackerApi.exceptions.BusTrackerException.BadRequest
 import busTrackerApi.exceptions.BusTrackerException.NotFound
 import busTrackerApi.extensions.getWrapped
 import busTrackerApi.routing.Response.ResponseJson
+import busTrackerApi.routing.lines.VehicleLocations
 import busTrackerApi.routing.lines.buildVehicleLocationJson
 import busTrackerApi.routing.stops.bus.busCodMode
 import busTrackerApi.utils.Pipeline
 import crtm.utils.createStopCode
 import crtm.utils.getCodModeFromLineCode
+import crtm.utils.getSimpleLineCodeFromLineCode
 import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cachingheaders.*
-import simpleJson.asJson
 
 suspend fun Pipeline.getLocations() = either {
     val lineCode = call.parameters.getWrapped("lineCode").bind()
@@ -31,12 +33,22 @@ suspend fun Pipeline.getLocations() = either {
         itinerary = getItinerariesResponse(lineCode).bind().firstOrNull { it.direction == direction - 1 }
             ?: shift<Nothing>(NotFound())
     }
+    val route = getRoute(lineCode).getOrNull()
+    val simpleLineCode = route?.simpleLineCode ?: getSimpleLineCodeFromLineCode(lineCode)
+    val routeCodMode = route?.codMode ?: busCodMode
 
-    val locations = getLocationsResponse(itinerary, lineCode, codMode, fullStopCode).bind()
+    val locations = VehicleLocations(
+        locations = getLocationsResponse(
+            itinerary,
+            lineCode,
+            codMode,
+            fullStopCode
+        ).bind().vehiclesLocation.vehicleLocation,
+        codMode = routeCodMode.toInt(),
+        lineCode = simpleLineCode,
+    )
 
-    val json = locations.vehiclesLocation.vehicleLocation
-        .map(::buildVehicleLocationJson)
-        .asJson()
+    val json = buildVehicleLocationJson(locations)
 
     call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 10))
     ResponseJson(json, HttpStatusCode.OK)