+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/emt/EmtLinesClient.kt
@@ -3,12 +3,14 @@ package busTrackerApi.routing.lines.emt
 import arrow.core.Either
 import arrow.core.continuations.either
 import busTrackerApi.db.getItineraryByFullLineCode
+import busTrackerApi.db.getRoute
 import busTrackerApi.exceptions.BusTrackerException
 import busTrackerApi.exceptions.BusTrackerException.BadRequest
 import busTrackerApi.exceptions.BusTrackerException.NotFound
 import busTrackerApi.extensions.getWrapped
 import busTrackerApi.routing.Response
 import busTrackerApi.routing.Response.ResponseJson
+import busTrackerApi.routing.lines.VehicleLocations
 import busTrackerApi.routing.lines.buildVehicleLocationJson
 import busTrackerApi.routing.stops.emt.emtCodMode
 import busTrackerApi.utils.Pipeline
@@ -18,7 +20,6 @@ import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cachingheaders.*
-import simpleJson.asJson
 
 suspend fun Pipeline.getLocations(): Either<BusTrackerException, Response> = either {
     val lineCode = call.parameters.getWrapped("lineCode").bind()
@@ -30,12 +31,19 @@ suspend fun Pipeline.getLocations(): Either<BusTrackerException, Response> = eit
         val itinerary = getItineraryByFullLineCode(lineCode, direction - 1) ?: shift<Nothing>(NotFound())
         itinerary.stops[itinerary.stops.size - 2].fullStopCode
     }
+    val route = getRoute(lineCode).getOrNull()
+    val simpleLineCode = route?.simpleLineCode ?: getSimpleLineCodeFromLineCode(lineCode)
+    val codMode = route?.codMode ?: emtCodMode
 
-    val locations =
-        getLocationsResponse(fullStopCode, getSimpleLineCodeFromLineCode(lineCode)).bind()
-            .filter { it.direction == direction }
-            .map(::buildVehicleLocationJson).asJson()
+    val locations = VehicleLocations(
+        locations = getLocationsResponse(fullStopCode, simpleLineCode).bind()
+            .filter { it.direction == direction },
+        codMode = codMode.toInt(),
+        lineCode = simpleLineCode
+    )
+
+    val json = buildVehicleLocationJson(locations)
 
     call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 15))
-    ResponseJson(locations, HttpStatusCode.OK)
+    ResponseJson(json, HttpStatusCode.OK)
 }